Update DSL version, make idempotent, move build
Update the DSL version to align with the TypeScript version.  Remove the
Travis CI token in favor of a more powerful GitHub token.  Move the
Travis modifications inside the project modifications so they only get
enacted if the initial predicate (directoryExists ".atomist") is true.

copyEditorBackingFileOrFail fails if the file already exists, so prepend
all calls to that function with a delete, which succeeds if the file is
not there.

Remove console log statements in TypeScript version.

Update build to use trusty, ensure a recent version of node is
installed, and run `npm install` if a package.json exists.  On Travis,
node versions are managed with `nvm`, which is a shell script that is
not exported, so you need to let Travis run it directly.  The Editor was
updated to reflect this requirement as well.

The output from the two different versions of the editor are slightly
different, but the end result is the same.  The TypeScript version
outputs a line for each travis.encrypt call while the Rug DSL version
outputs a single line showing the end results of all three encrypts.  In
both cases, all three encrypted variables are in the .travis.yml.

The changes this editor makes have not yet been committed to this repo.

Update README.